assignment/serializer/post_serializer.py
from django.contrib.auth.models import User
from rest_framework import serializers
from rest_framework_simplejwt.tokens import RefreshToken

from assignment.models import Post


class PostSerializer(serializers.ModelSerializer):

    class Meta:
        model = Post
        fields = ['id','user', 'text', 'created_at', 'updated_at']
        # fields = '__all__'


